Refactor plot events and write major throughline (#10)
* Refactor plot events and write major throughline

This includes nightly dreams and the climactic finales, even though it's not very climactic... still need exam

* Finished writing prologue plot scenes

Still need a bunch of gameplay screens to fill in the blanks - both in the exam and in the fire rescue.

Also need to add a bunch of character scenes. But it's coming together.